requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pandas>=2.2.0
plotly>=5.20.0
duckdb>=1.0.0
streamlit-sortables>=0.3.1
reportlab>=4.0.0
